tt_impl/tt rename; added tests to the CI
