+++ models/experimental/stable_diffusion_xl_base/tt/sdxl_utility.py
