+++ tests/nightly/single_card/stable_diffusion_xl_base/test_module_tt_crossattndownblock2d.py
@@ -1,0 +1,1 @@
+../../../../models/experimental/stable_diffusion_xl_base/tests/pcc/test_module_tt_crossattndownblock2d.py
